renamed copper_bar.rs to bars.rs
